+++ src/components/AfterLoginMain/NewStepWithWeMeet/NewStepWithWeMeet.scss
@@ -40,7 +40,7 @@
             padding: 0 16px 0 0;
 
             .background-green {
-                background-image: url(../../Images/motivation_green.png);
+                background-image: url(../../../Images/motivation_green.png);
                 height: 124px;
                 background-position: center center;
                 background-size: cover;
@@ -51,7 +51,7 @@
             }
 
             .background-yellow {
-                background-image: url(../../Images/motivation_yellow.png);
+                background-image: url(../../../Images/motivation_yellow.png);
                 height: 124px;
                 background-position: center center;
                 background-size: cover;
@@ -62,7 +62,7 @@
             }
 
             .background-red {
-                background-image: url(../../Images/motivation_red.png);
+                background-image: url(../../../Images/motivation_red.png);
                 height: 124px;
                 background-position: center center;
                 background-size: cover;
@@ -73,7 +73,7 @@
             }
 
             .background-blue {
-                background-image: url(../../Images/motivation_blue.png);
+                background-image: url(../../../Images/motivation_blue.png);
                 height: 124px;
                 background-position: center center;
                 background-size: cover;
