+++ src/components/AfterLoginMain/SimpleEventFilter/SimpleEventFilter.scss
